Move all api methods and classes into a core directory
