+++ dbcollection/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core API methods/classes.
+"""
